Updates hk2 plugin to 2.2.0
com.sun.mail.* Imports and Export have version 0.0.0 with hk2 2.0.0

Hk2 (a osgiversion-maven-plugin) is used to determine the version of
com.sun.mail.* Imports and Exports.
When built on different environments (and with different maven versions)
the results may vary and wrong import/export versions (like 0.0.0) are
created.

This issue should be fixed in a more recent version of the plugin, see:
https://java.net/jira/browse/HK2-125